Mac ambiguity fix

+++ src/mgl/configuration.cc
@@ -296,7 +296,7 @@ void GrueConfig::loadExtruderParams(const Configuration& config) {
     size_t extruderCount = config["extruderProfiles"].size();
     const Json::Value& extruderBase = config["extruderProfiles"];
     for (size_t i = 0; i < extruderCount; i++) {
-        const Json::Value &value = extruderBase[i];
+        const Json::Value &value = extruderBase[static_cast<int>(i)];
 
         stringstream ss;
         ss << "extruders[" << i << "].";
